LayoutTests/platform/chromium-win/plugins/windowless-paint-rect-coordinates-expected.txt

CONSOLE MESSAGE: PLUGIN: A null paint rect was passed in the WM_PAINT event

This tests that the paint rect passed in WM_PAINT has the expected coordinates.
